code cleanup: move old code to experiments module
