+++ experiments/README.md
@@ -1,0 +1,1 @@
+Unpolished code for exploring and experimenting with various ideas. Not intended for production use.
